[backend] refacto start investigation from report - move start investigation logic in investigation domain
