+++ opencti-platform/opencti-graphql/src/modules/workspace/investigation-domain.ts
@@ -8,6 +8,11 @@ import { ENTITY_TYPE_CONTAINER_REPORT } from '../../schema/stixDomainObject';
 import { STIX_SPEC_VERSION } from '../../database/stix';
 import { generateStandardId } from '../../schema/identifier';
 import type { StixId, StixObject } from '../../types/stix-common';
+import { BasicStoreEntityGrouping } from '../grouping/grouping-types';
+import { elList } from '../../database/middleware-loader';
+import { READ_INDEX_INTERNAL_OBJECTS } from '../../database/utils';
+import { addWorkspace } from './workspace-domain';
+import type { BasicStoreEntity } from '../../types/store';
 
 const EXPORTABLE_INVESTIGATED_TYPES: string[] = ['Stix-Core-Object', 'Stix-Core-Relationship', 'Stix-Sighting-Relationship'];
 
@@ -59,4 +64,44 @@ export const toStixReportBundle = async (context: AuthContext, user: AuthUser, w
   const bundle = buildStixReportBundle([stixInvestigationAsReportForExport, ...stixInvestigatedEntities]);
 
   return JSON.stringify(bundle);
+};
+
+const nameInvestigationToStartFromContainer = async (context: AuthContext, user: AuthUser, container: BasicStoreEntity) => {
+  const InvestigationToStartCanonicalName = `investigation from ${container.entity_type.toLowerCase()} "${container.name}"`;
+  const investigations: any = await elList(context, user, [READ_INDEX_INTERNAL_OBJECTS], {
+    filters: [{
+      key: 'entity_type',
+      values: ['workspace'],
+    }, {
+      key: 'type',
+      values: ['investigation']
+    }]
+  });
+
+  const investigationNameToMatch: RegExp = new RegExp(`^${InvestigationToStartCanonicalName} (\\d+)$`);
+  const highestInvestigationNumber = investigations
+    .map((investigation: { name: string; }) => {
+      const matches = investigation.name.match(investigationNameToMatch);
+
+      return matches ? Number(matches[1]) : 0;
+    })
+    .sort((a: number, b: number) => b - a)[0];
+
+  if (highestInvestigationNumber === undefined) {
+    return InvestigationToStartCanonicalName;
+  } if (highestInvestigationNumber === 0) {
+    return `${InvestigationToStartCanonicalName} 2`;
+  }
+
+  return `${InvestigationToStartCanonicalName} ${highestInvestigationNumber + 1}`;
+};
+
+export const startInvestigationFromContainer = async (context: AuthContext, user: AuthUser, container: BasicStoreEntity) => {
+  const investigationInput = {
+    type: 'investigation',
+    name: await nameInvestigationToStartFromContainer(context, user, container),
+    investigated_entities_ids: container.object
+  };
+
+  return addWorkspace(context, user, investigationInput);
 };
